type dependency에 의한 lint 오류 수정
