+++ src/components/common/card/InfoArea.tsx
@@ -7,7 +7,7 @@ import Position from './Position';
 import Typography from '../typography/Typography';
 
 import { Bottom, NameVoiceAndPosition, NameVoice, StyledInfoArea, Tier, Top } from './style';
-import { CardProps } from './LargeCard';
+import { CardProps } from '../../../types/card.type';
 
 type InfoAreaProps = Omit<CardProps, 'profileUrl'>;
