+++ src/components/common/card/LargeCard.tsx
@@ -2,6 +2,7 @@ import { useTheme } from '@emotion/react';
 import Image from 'next/image';
 import React from 'react';
 import { Typography } from '..';
+import { CardProps } from '../../../types/card.type';
 import PlayStyle from './PlayStyle';
 import Position from './Position';
 import {
@@ -13,19 +14,6 @@ import {
   RecommandInfoContainer,
   Tier,
 } from './style';
-import { Position as PositionType } from '../../../types/api.type';
-
-export interface CardProps {
-  leaguePoints: string;
-  lolNickname: string;
-  playStyle: string[];
-  position: PositionType[];
-  profileUrl: string;
-  rank: string;
-  tier: string;
-  useVoice: boolean;
-  _id?: string;
-}
 
 function LargeCard(props: CardProps) {
   const { profileUrl, tier, rank, lolNickname, position, playStyle, useVoice } = props;
